Display message type properly for MMS and SMS
When a message is showed details, android will search it in mms
& sms list according to its message id. But sometime the message
id of sms message is the same with that of mms message and the
special mms message is before the sms message in mms & sms list,
so the sms is displayed as mms.

To fix this bug, message type is to be checked at the same time.

Change-Id: I35f8586bf97a553a53424adb2458a8c7fd9ba0e9
CRs-Fixed: 568258
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>