MMS: Fix The slideshow can't be shown in the message
After the MessageItem Pdu Loaded, the MessageItem's AttachmentType
might have changed. So precisely, it's not the same MessageItem
anymore.Since it's not the same MessageItem, we should setImage,
and the slideshow will be shown normally.

After Pdu Loaded, when bindCommonMessage again, the sameItem
argument should be false, because it's not the same MessageItem
anymore.

CRs-Fixed: 548610

Change-Id: Id25bdfb1acc954d244ec3a8585fa1b0daa379d0f
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>